+++ Monitor/src/main/java/racecontrol/client/extension/racereport/EventRecord.java
@@ -12,13 +12,13 @@ import racecontrol.client.data.SessionId;
  * @author Leonard
  */
 public class EventRecord {
-    private final float sessionTime;
+    private final int sessionTime;
     private final String typeDesciptor;
     private final String info;
-    private final float replayTime;
+    private final int replayTime;
     private final SessionId sessionId;
 
-    public EventRecord(float sessionTime, String typeDesciptor, String info, float replayTime, SessionId sessionId) {
+    public EventRecord(int sessionTime, String typeDesciptor, String info, int replayTime, SessionId sessionId) {
         this.sessionTime = sessionTime;
         this.typeDesciptor = typeDesciptor;
         this.info = info;
@@ -26,7 +26,7 @@ public class EventRecord {
         this.sessionId = sessionId;
     }
 
-    public float getSessionTime() {
+    public int getSessionTime() {
         return sessionTime;
     }
 
@@ -38,7 +38,7 @@ public class EventRecord {
         return info;
     }
 
-    public float getReplayTime() {
+    public int getReplayTime() {
         return replayTime;
     }
